moved UnitMemento to own file
